+++ FirebaseMessaging/CHANGELOG.md
@@ -3,8 +3,8 @@
   delegates. (#15987)
 - [changed] **Breaking Change**: Scene delegates take priority over app delegates for
   automatic deep link routing. (#15987)
-- [fixed] Enforce `NSSecureCoding` when archiving and unarchiving registration
-  tokens in the keychain. (#16511)
+- [fixed] Fixed an issue where `NSSecureCoding` was not enforced when
+  archiving and unarchiving registration tokens in the keychain. (#16511)
 
 # 12.19.0
 - [fixed] Fix an issue where `messaging:didReceiveRegistrationToken:` was no